Create new branch, format all source code, fix url issue
